OMX-11: Separated elements in index.tsx into components
